+++ compiler/src/x86_arch_full.ml
@@ -119,6 +119,9 @@ module X86_core = struct
     | SBB _ -> true
     | SETcc -> true
     | SFENCE -> false (* Not DOIT *)
+    | SHA256MSG1 -> true
+    | SHA256MSG2 -> true
+    | SHA256RNDS2 -> true
     | SHL _ -> true
     | SHLD _ -> false (* Not DOIT *)
     | SHLX _ -> true